obid/requirements.txt
Django>=4.2,<7
djangorestframework
django-allauth
requests
PyJWT
cryptography
Pillow
